Fix icons and alarms (based on elsid's code).

+++ src/haven/Buff.java
@@ -119,7 +119,7 @@ public class Buff extends Widget implements ItemInfo.ResOwner, Bufflist.Managed 
 
     public void draw(GOut g) {
         g.chcolor(255, 255, 255, a);
-        Double ameter = (this.ameter >= 0) ? (this.ameter / 100.0) : ameteri.get();
+        Double ameter = (this.ameter >= 0) ? Double.valueOf(this.ameter / 100.0) : ameteri.get();
         if (ameter != null) {
             g.image(cframe, Coord.z);
             g.chcolor(0, 0, 0, a);
@@ -138,7 +138,7 @@ public class Buff extends Widget implements ItemInfo.ResOwner, Bufflist.Managed 
             Double cmeter;
             if (this.cmeter >= 0) {
                 double m = this.cmeter;
-                if(cmrem >= 0) {
+                if (cmrem >= 0) {
                     double ot = cmrem;
                     double pt = Utils.rtime() - gettime;
                     m *= (ot - pt) / ot;
@@ -206,17 +206,17 @@ public class Buff extends Widget implements ItemInfo.ResOwner, Bufflist.Managed 
     }
 
     public void reqdestroy() {
-	anims.clear();
-	final Coord o = this.c;
-	dest = true;
-	new NormAnim(0.35) {
-	    public void ntick(double a) {
-		Buff.this.a = 255 - (int)(255 * a);
-		Buff.this.c = o.add(0, (int)(a * a * cframe.sz().y));
-		if(a == 1.0)
-		    destroy();
-	    }
-	};
+        anims.clear();
+        final Coord o = this.c;
+        dest = true;
+        new NormAnim(0.35) {
+            public void ntick(double a) {
+                Buff.this.a = 255 - (int) (255 * a);
+                Buff.this.c = o.add(0, (int) (a * a * cframe.sz().y));
+                if (a == 1.0)
+                    destroy();
+            }
+        };
     }
 
     public void move(Coord c, double off) {
@@ -259,8 +259,8 @@ public class Buff extends Widget implements ItemInfo.ResOwner, Bufflist.Managed 
             this.nmeter = (Integer) args[0];
             ntext = null;
         } else if (msg == "cm") {
-            this.cmeter = ((Number)args[0]).doubleValue() / 100.0;
-            this.cmrem = (args.length > 1) ? (((Number)args[1]).doubleValue() * 0.06) : -1;
+            this.cmeter = ((Number) args[0]).doubleValue() / 100.0;
+            this.cmrem = (args.length > 1) ? (((Number) args[1]).doubleValue() * 0.06) : -1;
             gettime = Utils.rtime();
         } else {
             super.uimsg(msg, args);
